Add connections for SigNoz and Dynatrace (#62)

+++ src/ingester/api/handlers.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
+	"regexp"
 	"strings"
 
 	"ingester/auth"
@@ -42,6 +43,34 @@ func decodeRequestBody(r *http.Request, dest interface{}) error {
 	return json.NewDecoder(r.Body).Decode(dest)
 }
 
+func validateConnectionRequest(request *db.ConnectionRequest) error {
+	// Define valid platforms and their required fields
+	var err error
+
+	switch request.Platform {
+	case "grafana":
+		// For Grafana, all fields are required
+		if request.MetricsUsername == "" || request.LogsUserName == "" || request.ApiKey == "" || request.MetricsURL == "" || request.LogsURL == "" {
+			err = fmt.Errorf("valid parameters for platform 'grafana' are - 'metricsUsername', 'logsUserName', 'apiKey', 'metricsURL', and 'logsURL'")
+		}
+	case "datadog", "newrelic", "dynatrace":
+		// For Datadog, New Relic, and Dynatrace, apiKey, metricsURL, and logsURL are required
+		if request.ApiKey == "" || request.MetricsURL == "" || request.LogsURL == "" || request.MetricsUsername != "" || request.LogsUserName != "" {
+			err = fmt.Errorf("valid parameters for platform '%s' are - 'apiKey', 'metricsURL', and 'logsURL", request.Platform)
+		}
+	case "signoz":
+		// For SigNoz, apiKey and logsURL are required
+		if request.ApiKey == "" || request.LogsURL == "" {
+			err = fmt.Errorf("valid parameters for platform 'signoz' are - 'apiKey' and 'logsURL'")
+		}
+	default:
+		// Platform not supported
+		err = fmt.Errorf("unsupported platform '%s'. Supported platforms are: 'grafana', 'datadog', 'newrelic', 'dynatrace', 'signoz'", request.Platform)
+	}
+
+	return err
+}
+
 // getAuthKey retrieves the API Key from the request header.
 func getAuthKey(r *http.Request) string {
 	return r.Header.Get("Authorization")
@@ -77,6 +106,56 @@ func handleAPIKeyErrors(w http.ResponseWriter, err error, name string) {
 	}
 }
 
+// generateConnectionsHandler handles the creation of a new Connection
+func generateConnectionsHandler(w http.ResponseWriter, r *http.Request) {
+	var request db.ConnectionRequest
+
+	if err := decodeRequestBody(r, &request); err != nil {
+		sendJSONResponse(w, http.StatusBadRequest, errMsgInvalidBody)
+		return
+	}
+
+	if err := validateConnectionRequest(&request); err != nil {
+		sendJSONResponse(w, http.StatusBadRequest, err.Error())
+		return
+	}
+
+	err := db.GenerateConnection(getAuthKey(r), request)
+	if err != nil {
+		sendJSONResponse(w, http.StatusBadRequest, "Error creating connection: "+err.Error())
+		return
+	}
+
+	sendJSONResponse(w, http.StatusOK, "Connection created successfully")
+}
+
+// deleteConnectionsHandler handles the creation of a new Connection
+func deleteConnectionsHandler(w http.ResponseWriter, r *http.Request) {
+	var request db.ConnectionRequest
+
+	if err := decodeRequestBody(r, &request); err != nil {
+		sendJSONResponse(w, http.StatusBadRequest, errMsgInvalidBody)
+		return
+	}
+
+	if err := validateConnectionRequest(&request); err != nil {
+		sendJSONResponse(w, http.StatusBadRequest, err.Error())
+		return
+	}
+
+	err := db.DeleteConnection(getAuthKey(r))
+	if err != nil {
+		if err.Error() == "NOTFOUND" {
+			sendJSONResponse(w, http.StatusNotFound, "No existing Connection found")
+			return
+		}
+		sendJSONResponse(w, http.StatusBadRequest, "Error deleting connection: "+err.Error())
+		return
+	}
+
+	sendJSONResponse(w, http.StatusOK, "Connection deleted successfully")
+}
+
 // generateAPIKeyHandler handles the creation of a new API Key
 func generateAPIKeyHandler(w http.ResponseWriter, r *http.Request) {
 	var request APIKeyRequest
@@ -86,14 +165,17 @@ func generateAPIKeyHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	request.Normalize()
+	if request.Name != "doku-client-internal" {
+		newAPIKey, err := db.GenerateAPIKey(getAuthKey(r), request.Name)
+		if err != nil {
+			handleAPIKeyErrors(w, err, request.Name)
+			return
+		}
 
-	newAPIKey, err := db.GenerateAPIKey(getAuthKey(r), request.Name)
-	if err != nil {
-		handleAPIKeyErrors(w, err, request.Name)
-		return
+		sendJSONResponse(w, http.StatusOK, newAPIKey)
+	} else {
+		sendJSONResponse(w, http.StatusBadRequest, "API Key name 'doku-client-internal' is reserved and cannot be used")
 	}
-
-	sendJSONResponse(w, http.StatusOK, newAPIKey)
 }
 
 // getAPIKeyHandler handles retrieving an existing API key.
@@ -105,14 +187,17 @@ func getAPIKeyHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	request.Normalize()
+	if request.Name != "doku-client-internal" {
+		apiKey, err := db.GetAPIKeyForName(getAuthKey(r), request.Name)
+		if err != nil {
+			handleAPIKeyErrors(w, err, request.Name)
+			return
+		}
 
-	apiKey, err := db.GetAPIKeyForName(getAuthKey(r), request.Name)
-	if err != nil {
-		handleAPIKeyErrors(w, err, request.Name)
-		return
+		sendJSONResponse(w, http.StatusOK, apiKey)
+	} else {
+		sendJSONResponse(w, http.StatusBadRequest, "API Key name 'doku-client-internal' is reserved and cannot be accessed")
 	}
-
-	sendJSONResponse(w, http.StatusOK, apiKey)
 }
 
 // deleteAPIKeyHandler handles deleting an existing API key.
@@ -125,13 +210,16 @@ func deleteAPIKeyHandler(w http.ResponseWriter, r *http.Request) {
 	}
 
 	request.Normalize()
-	err := db.DeleteAPIKey(getAuthKey(r), request.Name)
-	if err != nil {
-		handleAPIKeyErrors(w, err, request.Name)
-		return
+	if request.Name != "doku-client-internal" {
+		err := db.DeleteAPIKey(getAuthKey(r), request.Name)
+		if err != nil {
+			handleAPIKeyErrors(w, err, request.Name)
+			return
+		}
+		sendJSONResponse(w, http.StatusOK, "API key deleted successfully")
+	} else {
+		sendJSONResponse(w, http.StatusBadRequest, "API Key name 'doku-client-internal' is reserved and cannot be deleted")
 	}
-
-	sendJSONResponse(w, http.StatusOK, "API key deleted successfully")
 }
 
 // DataHandler handles data related operations recieved on `/api/push` endpoint.
@@ -176,6 +264,46 @@ func APIKeyHandler(w http.ResponseWriter, r *http.Request) {
 	default:
 		w.WriteHeader(http.StatusMethodNotAllowed)
 	}
+}
+
+// ConnectionsHandler handles all 'Connections' tasks recieved on `/api/connections` endpoint.
+func ConnectionsHandler(w http.ResponseWriter, r *http.Request) {
+	switch r.Method {
+	case "POST":
+		generateConnectionsHandler(w, r)
+	case "DELETE":
+		deleteConnectionsHandler(w, r)
+	default:
+		w.WriteHeader(http.StatusMethodNotAllowed)
+	}
+}
+
+// retentionHandler handles updating the retention period in the Database.
+func RetentionHandler(w http.ResponseWriter, r *http.Request) {
+	type RetentionRequest struct {
+		RetentionPeriod string `json:"retentionPeriod"`
+	}
+	var request RetentionRequest
+
+	if err := decodeRequestBody(r, &request); err != nil {
+		sendJSONResponse(w, http.StatusBadRequest, errMsgInvalidBody)
+		return
+	}
+
+	// Use regex to validate the format "<int> days"
+	match, _ := regexp.MatchString(`^\d+\s+days$`, request.RetentionPeriod)
+	if !match {
+		sendJSONResponse(w, http.StatusBadRequest, "retention period should be in the format '<int> days'")
+		return
+	}
+
+	err := db.UpdateRetention(getAuthKey(r), request.RetentionPeriod)
+	if err != nil {
+		sendJSONResponse(w, http.StatusInternalServerError, err.Error())
+		return
+	}
+
+	sendJSONResponse(w, http.StatusOK, "retention period updated successfully")
 }
 
 // BaseEndpoint serves as a health check and entry point for the service.
